% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{Group}
\alias{Group}
\title{An abstract group.}
\usage{
Group(dimension, identity)
}
\arguments{
\item{dimension}{dimension of the group}

\item{identity}{the identity element of the group}
}
\description{
define a group

obtain a tensor containing all group elements in this group

define the group product on two elements

obtain inverse for a group element

group action of an element from the subgroup H on a vector in R2.
  For efficiency we implement this batchwise.

group action of elements of H on other elements in H itself. Comes down to group product.
  For efficiency we implement this batchwise. Each element in `h_batch` is applied to each element
  in `h_prime_batch`.

obtain a matrix representation in R2 for an element `h`

calculate the determinant of the representation of a group element `h`

map the group elements to an interval `[-1, 1]`. We use this to create
  a standardized input for obtaining weights over the group.
}
