% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupconv-kernel.R
\name{GroupConvKernel}
\alias{GroupConvKernel}
\title{Base functionality of a group convolution kernel}
\usage{
GroupConvKernel(group, kernel_size, in_channels, out_channels)
}
\arguments{
\item{group}{the group to use}

\item{kernel_size}{size of the convolution kernel}

\item{in_channels}{number of channels in the input}

\item{out_channels}{number of channels in the output layer}
}
\value{
a filter bank extending over all input channels, containing kernels
transformed for all output group elements.
}
\description{
define a group-to-group convolution kernel

Create a combined grid as the product of the transformed grids over R2 and H.

Sample convolution kernels for a given number of group elements
}
\details{
Kernel that maps between an incoming group and an outgoing group, both being semi-direct
products with R2.
}
