% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-convolution.R
\name{GroupConvolution}
\alias{GroupConvolution}
\title{Group convolution}
\usage{
GroupConvolution(group, kernel_size, in_channels, out_channels)
}
\arguments{
\item{group}{the group to use in lifting}

\item{kernel_size}{size of the convolution kernel}

\item{in_channels}{number of channels in the input}

\item{out_channels}{number of channels in the output layer}
}
\description{
implement a lifting convolution module
}
