% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-equivariant-cnn.R
\name{GroupEquivariantCNN}
\alias{GroupEquivariantCNN}
\title{Group equivariant CNN}
\usage{
GroupEquivariantCNN(
  group,
  kernel_size,
  in_channels,
  out_channels,
  num_hidden,
  hidden_channels
)
}
\arguments{
\item{group}{the group to use in lifting}

\item{kernel_size}{size of the convolution kernel}

\item{in_channels}{number of channels in the input}

\item{out_channels}{number of channels in the output layer}

\item{num_hidden}{number of hidden layers}

\item{hidden_channels}{number of channels in the hidden layers}
}
\description{
implement a lifting convolution module
}
