% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifting-kernel.R
\name{InterpolatingLiftingKernel}
\alias{InterpolatingLiftingKernel}
\title{A lifting kernel that does interpolation}
\usage{
InterpolatingLiftingKernel(group, kernel_size, in_channels, out_channels)
}
\arguments{
\item{group}{the group to use in lifting}

\item{kernel_size}{size of the convolution kernel}

\item{in_channels}{number of channels in the input}

\item{out_channels}{number of channels in the output layer}
}
\value{
a filter bank extending over all input channels, containing kernels
transformed for all output group elements.
}
\description{
define a lifting kernel

Sample convolution kernels for a given number of group elements
}
