% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifting-convolution.R
\name{LiftingConvolution}
\alias{LiftingConvolution}
\title{Lifting convolution}
\usage{
LiftingConvolution(group, kernel_size, in_channels, out_channels)
}
\arguments{
\item{group}{the group to use in lifting}

\item{kernel_size}{size of the convolution kernel}

\item{in_channels}{number of channels in the input}

\item{out_channels}{number of channels in the output layer}
}
\description{
implement a lifting convolution module
}
\details{
A convolution module taking as input a feature map defined over R2, and returning
a feature map over the semi-direct product of R2 and the symmetry group employed.
}
