% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifting-kernel.R
\name{LiftingKernel}
\alias{LiftingKernel}
\title{Base functionality of a lifting kernel}
\usage{
LiftingKernel(group, kernel_size, in_channels, out_channels)
}
\arguments{
\item{group}{the group to use in lifting}

\item{kernel_size}{size of the convolution kernel}

\item{in_channels}{number of channels in the input}

\item{out_channels}{number of channels in the output layer}

\item{sampled_group_elements}{the group elements over which to sample the convolution kernels}
}
\value{
a filter bank extending over all input channels, containing kernels
transformed for all output group elements.
}
\description{
define a lifting kernel

Transform the created grid by the group action of each group element.
This yields a grid (over H) of spatial grids (over R2). In other words,
a list of grids, each index of which is the original spatial grid transformed by
a corresponding group element in H.

Sample convolution kernels for a given number of group elements
}
\details{
The kernel lifts a signal from the input domain to the semi-direct product of the input domain
and the intended symmetry group.
}
